Move use-launch-your-store.js to ./hooks and update import path
