+++ plugins/woocommerce-admin/client/launch-your-store/hooks/use-launch-your-store.js
