+++ plugins/woocommerce-admin/client/launch-your-store/index.js
@@ -2,5 +2,5 @@
  * Internal dependencies
  */
 export { LaunchYourStoreStatus } from './status';
-export { useLaunchYourStore } from './use-launch-your-store';
+export { useLaunchYourStore } from './hooks/use-launch-your-store';
 export { registerSiteVisibilitySlotFill } from './settings/slotfill';
